some fixes

+++ python_modules/dagster-graphql/dagster_graphql/schema/pipelines/pipeline.py
@@ -282,7 +282,7 @@ class GrapheneAsset(graphene.ObjectType):
     definition = graphene.Field("dagster_graphql.schema.asset_graph.GrapheneAssetNode")
     latestEventSortKey = graphene.Field(graphene.ID)
     assetHealth = graphene.Field(GrapheneAssetHealth)
-    latestMaterializationTimestamp = graphene.Int()
+    latestMaterializationTimestamp = graphene.Float()
 
     class Meta:
         name = "Asset"
@@ -461,13 +461,13 @@ class GrapheneAsset(graphene.ObjectType):
 
     async def resolve_latestMaterializationTimestamp(
         self, graphene_info: ResolveInfo
-    ) -> Optional[int]:
+    ) -> Optional[float]:
         min_materialization_state = await MinimalAssetMaterializationHealthState.gen(
             graphene_info.context, self._asset_key
         )
         if min_materialization_state is not None:
             return (
-                int(min_materialization_state.latest_materialization_timestamp * 1000)
+                min_materialization_state.latest_materialization_timestamp * 1000
                 if min_materialization_state.latest_materialization_timestamp
                 else None
             )
@@ -475,9 +475,7 @@ class GrapheneAsset(graphene.ObjectType):
         record = await AssetRecord.gen(graphene_info.context, self._asset_key)
         latest_materialization_event = record.asset_entry.last_materialization if record else None
         return (
-            int(latest_materialization_event.timestamp * 1000)
-            if latest_materialization_event
-            else None
+            latest_materialization_event.timestamp * 1000 if latest_materialization_event else None
         )
